async FileIO class added
