+++ cve_bin_tool/extractor.py
@@ -11,7 +11,7 @@ import sys
 import tempfile
 
 from .log import LOGGER
-from .util import (
+from .async_utils import (
     aio_unpack_archive,
     aio_run_command,
     aio_glob,
@@ -19,8 +19,8 @@ from .util import (
     aio_rmdir,
     aio_makedirs,
     aio_mkdtemp,
-    file_writer,
     ChangeDirContext,
+    FileIO,
 )
 
 
@@ -77,7 +77,8 @@ class BaseExtractor(object):
                 if stderr or not stdout:
                     return 1
                 cpio_path = os.path.join(extraction_path, "data.cpio")
-                await file_writer(cpio_path, "wb", stdout)
+                async with FileIO(cpio_path, "wb") as f:
+                    await f.write(stdout)
                 stdout, stderr = await aio_run_command(
                     ["cpio", "-idm", "--file", cpio_path]
                 )
