+++ cve_bin_tool/util.py
@@ -1,50 +1,8 @@
 # pylint: disable=too-many-arguments
 """ Utility classes for the CVE Binary Tool """
-import asyncio
 import fnmatch
-import glob
 import os
-import shutil
 import sys
-import tempfile
-from functools import partial, wraps
-
-
-def async_wrap(func):
-    @wraps(func)
-    async def run(*args, loop=None, executor=None, **kwargs):
-        if loop is None:
-            loop = asyncio.get_event_loop()
-        pfunc = partial(func, *args, **kwargs)
-        return await loop.run_in_executor(executor, pfunc)
-
-    return run
-
-
-async def aio_run_command(args):
-    process = await asyncio.create_subprocess_exec(
-        *args, stdout=asyncio.subprocess.PIPE, stderr=asyncio.subprocess.PIPE,
-    )
-    stdout, stderr = await process.communicate()
-    return stdout, stderr  # binary encoded
-
-
-@async_wrap
-def file_writer(name, mode, content):
-    with open(name, mode) as f:
-        f.write(content)
-
-
-class ChangeDirContext:
-    def __init__(self, destination_dir):
-        self.current_dir = os.getcwd()
-        self.destination_dir = destination_dir
-
-    async def __aenter__(self):
-        os.chdir(self.destination_dir)
-
-    async def __aexit__(self, exc_type, exc_val, exc_tb):
-        os.chdir(self.current_dir)
 
 
 def regex_find(lines, version_patterns):
@@ -151,12 +109,3 @@ class DirWalk:
             if fnmatch.fnmatch(text, pattern):
                 return True
         return False
-
-
-aio_rmdir = async_wrap(shutil.rmtree)
-aio_rmfile = async_wrap(os.remove)
-aio_unpack_archive = async_wrap(shutil.unpack_archive)
-aio_glob = async_wrap(glob.glob)
-aio_mkdtemp = async_wrap(tempfile.mkdtemp)
-aio_makedirs = async_wrap(os.makedirs)
-aio_inpath = async_wrap(inpath)
